+++ library/src/main/java/com/dengjinwen/basetool/library/tool/UnitConversionTool.java
@@ -12,7 +12,7 @@ public class UnitConversionTool {
      * @param dpValue
      * @return
      */
-    public static int dip2px(Context context,float dpValue){
+    public static float dip2px(Context context,float dpValue){
         return dp2px(context,dpValue);
     }
 
@@ -22,9 +22,9 @@ public class UnitConversionTool {
      * @param dpValue
      * @return
      */
-    public static int dp2px(Context context,float dpValue){
+    public static float dp2px(Context context,float dpValue){
         final float scale=context.getResources().getDisplayMetrics().density;
-        return (int)(dpValue*scale+0.5f);
+        return dpValue*scale+0.5f;
     }
 
     /**
@@ -33,7 +33,7 @@ public class UnitConversionTool {
      * @param pxValue
      * @return
      */
-    public static int px2dip(Context context, float pxValue){
+    public static float px2dip(Context context, float pxValue){
         return px2dp(context,pxValue);
     }
 
@@ -43,9 +43,9 @@ public class UnitConversionTool {
      * @param pxValue
      * @return
      */
-    public static int px2dp(Context context,float pxValue){
+    public static float px2dp(Context context,float pxValue){
         final float scale=context.getResources().getDisplayMetrics().density;
-        return (int) (pxValue/scale+0.5f);
+        return pxValue/scale+0.5f;
     }
 
     /**
@@ -54,9 +54,9 @@ public class UnitConversionTool {
      * @param spValue
      * @return
      */
-    public static int sp2px(Context context,float spValue){
+    public static float sp2px(Context context,float spValue){
         final float fontScale=context.getResources().getDisplayMetrics().scaledDensity;
-        return (int) (spValue*fontScale+0.5f);
+        return spValue*fontScale+0.5f;
     }
 
     /**
@@ -65,8 +65,8 @@ public class UnitConversionTool {
      * @param pxValue
      * @return
      */
-    public static int px2sp(Context context,float pxValue){
+    public static float px2sp(Context context,float pxValue){
         final float fontScale=context.getResources().getDisplayMetrics().scaledDensity;
-        return (int) (pxValue/fontScale+0.5f);
+        return pxValue/fontScale+0.5f;
     }
 }
